+++ include/mufevent.h
@@ -16,7 +16,7 @@ struct frame *muf_event_pid_frame(int pid);
 void muf_event_process(void);
 void muf_event_purge(struct frame *fr);
 int muf_event_read_notify(int descr, dbref player, const char *cmd);
-void muf_event_register_specific(dbref player, dbref prog, struct frame *fr, int eventcount,
+void muf_event_register_specific(dbref player, dbref prog, struct frame *fr, size_t eventcount,
 				 char **eventids);
 
 #endif				/* _MUFEVENT_H */
